+++ custom_components/better_thermostat/events/window.py
@@ -39,6 +39,10 @@ async def trigger_window_change(self, event) -> None:
                 "better_thermostat %s: Window sensor state is unknown, assuming window is open",
                 self.name,
             )
+        
+        # window was opened, disable heating power calculation for this period
+        self.heating_start_temp = None
+        self.async_write_ha_state()
     elif new_state == "off":
         new_window_open = False
     else:
